chore: change fixture name docs: add README
