+++ README.md
@@ -1,21 +1,8 @@
-# React Zeroconfig
+# 🚀 Rocket Scripts
 
-[![NPM](https://img.shields.io/npm/v/react-zeroconfig.svg)](https://www.npmjs.com/package/react-zeroconfig)
-[![Build Status](https://travis-ci.com/react-zeroconfig/react-zeroconfig.svg?branch=master)](https://travis-ci.com/react-zeroconfig/react-zeroconfig)
-[![DeepScan grade](https://deepscan.io/api/teams/3270/projects/5786/branches/142107/badge/grade.svg)](https://deepscan.io/dashboard#view=project&tid=3270&pid=5786&bid=142107)
-[![Coverage Status](https://coveralls.io/repos/github/react-zeroconfig/react-zeroconfig/badge.svg)](https://coveralls.io/github/react-zeroconfig/react-zeroconfig)
-[![Total alerts](https://img.shields.io/lgtm/alerts/g/react-zeroconfig/react-zeroconfig.svg?logo=lgtm&logoWidth=18)](https://lgtm.com/projects/g/react-zeroconfig/react-zeroconfig/alerts/)
-
-# 🚀 Quick Start
-
-- [✈️ Click to Create a Javascript Web App](https://github.com/react-zeroconfig/basic-javascript-app/generate)
-- [✈️ Click to Create a Javascript Electron Desktop App](https://github.com/react-zeroconfig/basic-javascript-electron-app/generate)
-- [✈️ Click to Create a Javascript SSR Web App](https://github.com/react-zeroconfig/basic-javascript-ssr-app/generate)
-- [✈️ Click to Create a Javascript Playground (storybook, jest...)](https://github.com/react-zeroconfig/javascript-playground/generate)
-- [🚀 Click to Create a Typescript Web App](https://github.com/react-zeroconfig/basic-typescript-app/generate)
-- [🚀 Click to Create a Typescript Electron Desktop App](https://github.com/react-zeroconfig/basic-typescript-electron-app/generate)
-- [🚀 Click to Create a Typescript SSR Web App](https://github.com/react-zeroconfig/basic-typescript-ssr-app/generate)
-- [🚀 Click to Create a Typescript Playground (storybook, jest...)](https://github.com/react-zeroconfig/typescript-playground/generate)
+[![NPM](https://img.shields.io/npm/v/rocket-scripts.svg)](https://www.npmjs.com/package/rocket-scripts)
+<!-- TODO CI badge -->
+<!-- Coveralls badge -->
 
 # License
-`react-zeroconfig` is open source software licensed as MIT.
+`rocket-scripts` is open source software licensed as MIT.
