+++ test/fixtures/web/proxy/public/favicon.ico
