+++ test/fixtures/web/proxy/public/manifest.json
