+++ test/fixtures/web/proxy/src/typings.d.ts
